Remove unused layer of abstraction from Settings
The data structures that represent compiler settings were refactored
to enable an immutable version to complement, and presumable replace,
the current mutable implementation. However we never created an
immutable version.

This commit removes that layer of abstraction, which isn't paying
its way.
